requirements.txt
Cython
pylint
six
gitpython
cython
python-dotenv
google-generativeai
fastapi
uvicorn[standard]
